+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [1.6.3-SNAPSHOT]
 
+- JSON `datetime` function has optional `millis` parameter to convert to/from milliseconds since epoch instead of seconds
+- `JsonConfig.DefaultDateTimeMillis = true` will cause `datetime` to convert to/from milliseconds since epoch by default
 - Improved Form prefWidth calculations
 - MenuItem.enableWhen function
 - Custom tab support for Views. Views can be docked in tabs and even delegate to refreshable and savable for the surrounding View